+++ importers/src/Utils/CoverVendor/VendorImageItem.php
@@ -16,7 +16,18 @@ class VendorImageItem implements \Stringable
     private Vendor $vendor;
     private string $originalFile;
     private \DateTime $originalLastModified;
-    private int $originalContentLength;
+    private int $originalContentLength = 0;
+    private string $eTag;
+
+    /**
+     * @param Vendor $vendor
+     * @param string $originalFile
+     */
+    public function __construct(string $originalFile, Vendor $vendor)
+    {
+        $this->originalFile = $originalFile;
+        $this->vendor = $vendor;
+    }
 
     public function __toString(): string
     {
@@ -116,9 +127,9 @@ class VendorImageItem implements \Stringable
     }
 
     /**
-     * @return mixed
+     * @return int
      */
-    public function getOriginalContentLength(): ?int
+    public function getOriginalContentLength(): int
     {
         return $this->originalContentLength;
     }
@@ -130,7 +141,27 @@ class VendorImageItem implements \Stringable
      */
     public function setOriginalContentLength(?int $originalContentLength): self
     {
-        $this->originalContentLength = $originalContentLength;
+        if (null !== $originalContentLength) {
+            $this->originalContentLength = $originalContentLength;
+        }
+
+        return $this;
+    }
+
+    /**
+     * @return string|null
+     */
+    public function getETag(): ?string
+    {
+        return $this->eTag;
+    }
+
+    /**
+     * @return static
+     */
+    public function setETag(string $eTag): self
+    {
+        $this->eTag = $eTag;
 
         return $this;
     }