+++ importers/tests/Service/VendorService/VendorImageValidatorServiceTest.php
@@ -8,9 +8,12 @@
 namespace Tests\Service\VendorService;
 
 use App\Entity\Source;
+use App\Entity\Vendor;
+use App\Service\VendorService\VendorImageDefaultValidator;
 use App\Service\VendorService\VendorImageValidatorService;
 use App\Utils\CoverVendor\VendorImageItem;
 use PHPUnit\Framework\TestCase;
+use Psr\Log\LoggerInterface;
 use Symfony\Component\HttpClient\MockHttpClient;
 use Symfony\Component\HttpClient\Response\MockResponse;
 
@@ -35,10 +38,14 @@ class VendorImageValidatorServiceTest extends TestCase
             ]),
         ]);
 
-        $item = new VendorImageItem();
-        $item->setOriginalFile($this->url);
+        $logger = $this->createMock(LoggerInterface::class);
 
-        $service = new VendorImageValidatorService($client);
+        $vendor = $this->createMock(Vendor::class);
+        $vendor->setId(1);
+        $item = new VendorImageItem($this->url, $vendor);
+
+        $defaultValidator = new VendorImageDefaultValidator($client, $logger);
+        $service = new VendorImageValidatorService($defaultValidator, []);
         $service->validateRemoteImage($item);
 
         $this->assertEquals(true, $item->isFound());
@@ -59,10 +66,15 @@ class VendorImageValidatorServiceTest extends TestCase
             ]),
         ]);
 
-        $item = new VendorImageItem();
+        $logger = $this->createMock(LoggerInterface::class);
+
+        $vendor = $this->createMock(Vendor::class);
+        $vendor->setId(1);
+        $item = new VendorImageItem($this->url, $vendor);
         $item->setOriginalFile($this->url);
 
-        $service = new VendorImageValidatorService($client);
+        $defaultValidator = new VendorImageDefaultValidator($client, $logger);
+        $service = new VendorImageValidatorService($defaultValidator, []);
         $service->validateRemoteImage($item);
 
         $this->assertEquals(false, $item->isFound());
@@ -83,10 +95,14 @@ class VendorImageValidatorServiceTest extends TestCase
             ]),
         ]);
 
+        $logger = $this->createMock(LoggerInterface::class);
+
         $timezone = new \DateTimeZone('UTC');
         $lastModifiedDateTime = \DateTime::createFromFormat('D, d M Y H:i:s \G\M\T', $this->lastModified, $timezone);
 
-        $item = new VendorImageItem();
+        $vendor = $this->createMock(Vendor::class);
+        $vendor->setId(1);
+        $item = new VendorImageItem($this->url, $vendor);
         $item->setOriginalFile($this->url)
             ->setOriginalContentLength($this->contentLength)
             ->setOriginalLastModified($lastModifiedDateTime);
@@ -96,7 +112,8 @@ class VendorImageValidatorServiceTest extends TestCase
             ->setOriginalContentLength($this->contentLength)
             ->setOriginalLastModified($lastModifiedDateTime);
 
-        $service = new VendorImageValidatorService($client);
+        $defaultValidator = new VendorImageDefaultValidator($client, $logger);
+        $service = new VendorImageValidatorService($defaultValidator, []);
         $service->isRemoteImageUpdated($item, $source);
 
         $this->assertEquals(false, $item->isUpdated());
@@ -117,10 +134,14 @@ class VendorImageValidatorServiceTest extends TestCase
             ]),
         ]);
 
+        $logger = $this->createMock(LoggerInterface::class);
+
         $timezone = new \DateTimeZone('UTC');
         $lastModifiedDateTime = \DateTime::createFromFormat('D, d M Y H:i:s \G\M\T', $this->lastModified, $timezone);
 
-        $item = new VendorImageItem();
+        $vendor = $this->createMock(Vendor::class);
+        $vendor->setId(1);
+        $item = new VendorImageItem($this->url, $vendor);
         $item->setOriginalFile($this->url)
             ->setOriginalContentLength($this->contentLength)
             ->setOriginalLastModified($lastModifiedDateTime);
@@ -130,30 +151,11 @@ class VendorImageValidatorServiceTest extends TestCase
             ->setOriginalContentLength($this->contentLength + 200)
             ->setOriginalLastModified($lastModifiedDateTime);
 
-        $service = new VendorImageValidatorService($client);
+        $defaultValidator = new VendorImageDefaultValidator($client, $logger);
+        $service = new VendorImageValidatorService($defaultValidator, []);
         $service->isRemoteImageUpdated($item, $source);
 
         $this->assertEquals(true, $item->isFound());
         $this->assertEquals(true, $item->isUpdated());
-    }
-
-    /**
-     * Test remoteImageHeader parser.
-     */
-    public function testRemoteImageHeader()
-    {
-        $client = new MockHttpClient([
-            new MockResponse('', [
-                'http_code' => 200,
-                'response_headers' => [
-                    'cf-polished' => 'origFmt=png, origSize=25272',
-                ],
-            ]),
-        ]);
-
-        $service = new VendorImageValidatorService($client);
-        $headers = $service->remoteImageHeader('cf-polished', $this->url);
-
-        $this->assertEquals(['origFmt=png, origSize=25272'], $headers);
     }
 }